[CCUBE-1524][MAHI]Cleaned up the exports for public helper functions
