+++ src/theme/colour-primitive/types.ts
@@ -1,7 +1,100 @@
-import { ColourScheme, PrimitiveColourSet } from "../types";
+import { ColourScheme } from "../types";
 
 export type ColourCollectionsMap = {
     [key in ColourScheme]: PrimitiveColourSet;
 };
 
 export type PrimitiveColourSetOptions = Partial<PrimitiveColourSet>;
+
+export type PrimitiveColourSet = {
+    "brand-10": string;
+    "brand-20": string;
+    "brand-30": string;
+    "brand-40": string;
+    "brand-50": string;
+    "brand-60": string;
+    "brand-70": string;
+    "brand-80": string;
+    "brand-90": string;
+    "brand-95": string;
+    "brand-100": string;
+    "primary-10": string;
+    "primary-20": string;
+    "primary-30": string;
+    "primary-40": string;
+    "primary-50": string;
+    "primary-60": string;
+    "primary-70": string;
+    "primary-80": string;
+    "primary-90": string;
+    "primary-95": string;
+    "primary-100": string;
+    "secondary-10": string;
+    "secondary-20": string;
+    "secondary-30": string;
+    "secondary-40": string;
+    "secondary-50": string;
+    "secondary-60": string;
+    "secondary-70": string;
+    "secondary-80": string;
+    "secondary-90": string;
+    "secondary-95": string;
+    "secondary-100": string;
+    "neutral-10": string;
+    "neutral-20": string;
+    "neutral-30": string;
+    "neutral-40": string;
+    "neutral-50": string;
+    "neutral-60": string;
+    "neutral-70": string;
+    "neutral-80": string;
+    "neutral-90": string;
+    "neutral-95": string;
+    "neutral-100": string;
+    "success-10": string;
+    "success-20": string;
+    "success-30": string;
+    "success-40": string;
+    "success-50": string;
+    "success-60": string;
+    "success-70": string;
+    "success-80": string;
+    "success-90": string;
+    "success-95": string;
+    "success-100": string;
+    "warning-10": string;
+    "warning-20": string;
+    "warning-30": string;
+    "warning-40": string;
+    "warning-50": string;
+    "warning-60": string;
+    "warning-70": string;
+    "warning-80": string;
+    "warning-90": string;
+    "warning-95": string;
+    "warning-100": string;
+    "error-10": string;
+    "error-20": string;
+    "error-30": string;
+    "error-40": string;
+    "error-50": string;
+    "error-60": string;
+    "error-70": string;
+    "error-80": string;
+    "error-90": string;
+    "error-95": string;
+    "error-100": string;
+    "info-10": string;
+    "info-20": string;
+    "info-30": string;
+    "info-40": string;
+    "info-50": string;
+    "info-60": string;
+    "info-70": string;
+    "info-80": string;
+    "info-90": string;
+    "info-95": string;
+    "info-100": string;
+    white: string;
+    black: string;
+};
